[REFACTOR] Move user dashboard stages into folder
