+++ src/client/react/Root.js
@@ -6,10 +6,10 @@ import { loadState } from '../actions/stateActions';
 import LoadingPage from './pages/Loading';
 import { getPublicSetting } from '../graphql/setting';
 import AppContainer from '../../../lib/react/components/AppContainer';
-import Promotion from './user/Promotion';
-import Registration from './user/Registration';
-import RegistrationClosed from './user/RegistrationClosed';
-import Race from './user/Race';
+import Promotion from './user/stages/Promotion';
+import Registration from './user/stages/Registration';
+import RegistrationClosed from './user/stages/RegistrationClosed';
+import Race from './user/stages/Race';
 
 import './scss/main.scss';
 import '../assets/favicon.ico';
