+++ src/client/react/user/stages/Race/Dashboard.js
@@ -2,13 +2,13 @@ import React from 'react';
 import { Route } from 'react-router-dom';
 import { connect } from 'react-redux';
 import { Switch, Redirect } from 'react-router-dom';
-import Base from '../components/Base';
-import HelpMenu from '../components/HelpMenu';
-import Home from '../views/Home';
-import Help from '../views/Help';
-import Instructions from '../views/Instructions';
-import ImageUploaderTest from '../views/ImageUploaderTest';
-import NotFound from '../views/NotFound';
+import Base from '../../components/Base';
+import HelpMenu from '../../components/HelpMenu';
+import Home from '../../views/Home';
+import Help from '../../views/Help';
+import Instructions from '../../views/Instructions';
+import ImageUploaderTest from '../../views/ImageUploaderTest';
+import NotFound from '../../views/NotFound';
 
 
 const mapStateToProps = (state, ownProps) => {
