+++ src/client/react/user/stages/Race/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { BrowserRouter, Switch, Route, Redirect } from 'react-router-dom';
-import NotFoundPage from '../../pages/NotFound';
-import AppContainer from '../../../../../lib/react/components/AppContainer';
-import Login from '../components/Login';
+import NotFoundPage from '../../../pages/NotFound';
+import AppContainer from '../../../../../../lib/react/components/AppContainer';
+import Login from '../../components/Login';
 import Dashboard from './Dashboard';
