+++ src/client/react/user/stages/Registration/RegisterForm/Form2.js
@@ -2,10 +2,10 @@ import React from 'react';
 import autobind from 'core-decorators/es/autobind';
 import { withApollo, gql } from 'react-apollo';
 import { Button, Intent } from '@blueprintjs/core';
-import FormInput from '../../../../../../lib/react/components/forms/FormInput';
-import Validator from '../../../../../../lib/react/components/forms/validation/Validator';
-import NotEmpty from '../../../../../../lib/react/components/forms/validation/functions/NotEmpty';
-import RegexCheck from '../../../../../../lib/react/components/forms/validation/functions/RegexCheck';
+import FormInput from '../../../../../../../lib/react/components/forms/FormInput';
+import Validator from '../../../../../../../lib/react/components/forms/validation/Validator';
+import NotEmpty from '../../../../../../../lib/react/components/forms/validation/functions/NotEmpty';
+import RegexCheck from '../../../../../../../lib/react/components/forms/validation/functions/RegexCheck';
 import { errorProps, pendingProps, successProps } from './lib';
