+++ src/client/react/user/stages/Registration/RegisterForm/index.js
@@ -4,8 +4,8 @@ import autobind from 'core-decorators/es/autobind';
 import { connect } from 'react-redux';
 import { Link, Redirect } from 'react-router-dom';
 import { ProgressBar, Intent, Icon } from '@blueprintjs/core';
-import { logout } from '../../../../actions/authActions';
-import API from '../../../../API';
+import { logout } from '../../../../../actions/authActions';
+import API from '../../../../../API';
 import Form1 from './Form1';
 import Form2 from './Form2';
 import Form3 from './Form3';
