+++ src/client/react/user/stages/Registration/RegisterForm/lib.js
