+++ src/client/react/user/stages/Registration/index.js
@@ -1,14 +1,15 @@
 import React from 'react';
 import { BrowserRouter, Switch, Route, Redirect } from 'react-router-dom';
-import NotFoundPage from '../../pages/NotFound';
-import AppContainer from '../../../../../lib/react/components/AppContainer';
-import Login from '../components/Login';
-import Pay from '../components/Pay';
+import NotFoundPage from '../../../pages/NotFound';
+import AppContainer from '../../../../../../lib/react/components/AppContainer';
+import Login from '../../components/Login';
+import Pay from '../../components/Pay';
 import Home from './Home';
+import Register from './Register';
 import Dashboard from './Dashboard';
 
 
-class RegistrationClosed extends React.Component {
+class Registration extends React.Component {
 	render() {
 		return (
 			<BrowserRouter>
@@ -16,19 +17,19 @@ class RegistrationClosed extends React.Component {
 					<Switch>
 						
 						<Route exact path='/' component={Home}/>
+						<Route exact path='/register' component={Register}/>
 						<Route exact path='/login' component={Login}/>
 						<Route exact path='/pay' component={Pay}/>
-						<Route path='/dashboard' component={Dashboard}/>
+						<Route exact path='/dashboard' component={Dashboard}/>
 
 						<Route component={NotFoundPage}/>
 
 					</Switch>
 				</AppContainer>
 			</BrowserRouter>
-
 		);
 	}
 }
 
 
-export default RegistrationClosed;
+export default Registration;
