+++ src/client/react/user/stages/RegistrationClosed/Dashboard.js
@@ -2,12 +2,12 @@ import React from 'react';
 import { Route } from 'react-router-dom';
 import { connect } from 'react-redux';
 import { Switch, Redirect } from 'react-router-dom';
-import Base from '../components/Base';
-import HelpMenu from '../components/HelpMenu';
-import Home from '../views/Home';
-import Help from '../views/Help';
-import Instructions from '../views/Instructions';
-import NotFound from '../views/NotFound';
+import Base from '../../components/Base';
+import HelpMenu from '../../components/HelpMenu';
+import Home from '../../views/Home';
+import Help from '../../views/Help';
+import Instructions from '../../views/Instructions';
+import NotFound from '../../views/NotFound';
 
 
 const mapStateToProps = (state, ownProps) => {
